Changed profile layout, added change password layout, started redesigning look of layouts
