+++ Android/app/src/main/res/drawable/rounded_corners.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <corners android:radius="24dp" />
+    <corners android:radius="10dp" />
     <solid android:color="#FFFFFF" />
 </shape>
