+++ Android/app/src/main/res/values-es/strings.xml
@@ -5,14 +5,18 @@
     <string name="addPhoto"> Añadir imagen</string>
     <string name="addPhotoDescription">Añade fotos de tu galería o haz una foto con tu cámara</string>
     <string name="appName">Janari0</string>
+    <string name="confirm">confirmar</string>
     <string name="contactSeller">Contactar con el vendedor</string>
     <string name="createAccount">Crear cuenta</string>
     <string name="defaultDescription"> Sin descripción </string>
+    <string name="defaultEmail">name.lastname@gmail.com</string>
+    <string name="defaultName">John Doe</string>
     <string name="defaultUploadedDescription">164389594560285.jpg</string>
     <string name="description">Descripción</string>
     <string name="donate">Donar</string>
     <string name="dummy"> Holly smokes, it works!</string>
     <string name="email">Email</string>
+    <string name="email2">email@email.com</string>
     <string name="enterDescription">Introduzca la descripción:</string>
     <string name="enterExpirationDate">Introduzca la fecha de caducidad:</string>
     <string name="enterNameProduct">Introduzca el nombre del producto:</string>
@@ -22,10 +26,14 @@
     <string name="forgetPassword">¿Ha olvidado su contraseña?</string>
     <string name="fromGallery">De la galería</string>
     <string name="incorrectPassword">Nombre de usuario o contraseña incorrectos</string>
+    <string name="layoutChangePassword">Cambiar la contraseña</string>
+    <string name="layoutProfile">Perfil</string>
     <string name="login">Inicio de sesión</string>
     <string name="name">Nombre</string>
+    <string name="newPasword">Nueva Contraseña</string>
     <string name="next">Siguiente</string>
     <string name="numberProducts"> 3</string>
+    <string name="oldPasword">Antigua Contraseña</string>
     <string name="onSale">EN VENTA!</string>
     <string name="oneWeek"> DENTRO DE UNAS SEMANAS</string>
     <string name="password">Constraseña</string>
@@ -38,15 +46,10 @@
     <string name="searchProduct"> buscar un producto</string>
     <string name="selectProduct"> Seleccionar producto</string>
     <string name="sellIt">Vender!</string>
+    <string name="signOut">Desconectar</string>
     <string name="submit"> Enviar</string>
     <string name="takeAPhoto">Toma una foto</string>
     <string name="user">Usuario</string>
     <string name="username">Nombre de usuario</string>
     <string name="welcome">Bienvenido</string>
-    <string name="oldPasword">Antigua Contraseña</string>
-    <string name="newPasword">Nueva Contraseña</string>
-    <string name="confirm">confirmar</string>
-    <string name="email2">email@email.com</string>
-    <string name="signOut">Desconectar</string>
-    <string name="defaultName">John Doe</string>
 </resources>
