+++ Android/app/src/main/res/values/strings.xml
@@ -5,14 +5,18 @@
     <string name="addPhoto"> Add photo</string>
     <string name="addPhotoDescription">Add photos from your gallery or take a picture with your camera</string>
     <string name="appName">Janari0</string>
+    <string name="confirm">confirm</string>
     <string name="contactSeller">Contact Seller</string>
     <string name="createAccount">Create account</string>
     <string name="defaultDescription">No description</string>
+    <string name="defaultEmail">name.lastname@gmail.com</string>
+    <string name="defaultName">Scofield11</string>
     <string name="defaultUploadedDescription">164389594560285.jpg</string>
     <string name="description">Description</string>
     <string name="donate">Donate</string>
     <string name="dummy"> Holly smokes, it works!</string>
     <string name="email">Email</string>
+    <string name="email2">email@email.com</string>
     <string name="enterDescription">Enter description:</string>
     <string name="enterExpirationDate">Enter expiration date:</string>
     <string name="enterNameProduct">Enter name of the product:</string>
@@ -22,10 +26,14 @@
     <string name="forgetPassword">Forgot your password?</string>
     <string name="fromGallery">From gallery</string>
     <string name="incorrectPassword">Incorrect username or password</string>
-    <string name="login">Login</string>
+    <string name="layoutChangePassword">Change password</string>
+    <string name="layoutProfile">Profile</string>
+    <string name="login">Login</string>gi9
     <string name="name">Name</string>
+    <string name="newPasword">New Password</string>
     <string name="next">Next</string>
     <string name="numberProducts"> 3</string>
+    <string name="oldPasword">Old Password</string>
     <string name="onSale"> ON SALE!</string>
     <string name="oneWeek"> WITHIN ONE WEEK</string>
     <string name="password">Password</string>
@@ -38,15 +46,10 @@
     <string name="searchProduct"> Search a product</string>
     <string name="selectProduct"> Select Product</string>
     <string name="sellIt">Sell It!</string>
+    <string name="signOut">Sign Out</string>
     <string name="submit">Submit</string>
     <string name="takeAPhoto">Take a photo</string>
     <string name="user">User</string>
     <string name="username">Username</string>
     <string name="welcome">Welcome</string>
-    <string name="oldPasword">Old Password</string>
-    <string name="newPasword">New Password</string>
-    <string name="confirm">confirm</string>
-    <string name="email2">email@email.com</string>
-    <string name="signOut">Sign Out</string>
-    <string name="defaultName">John Doe</string>
 </resources>
